Tests/LibWeb/Text/expected/css/shadow-host-animation-name-uses-shadow-root-keyframes.txt

opacity: 0.25
